+++ src/class-type.ts
@@ -1,4 +1,5 @@
-import { AnyObject } from "./interfaces";
+import { AnyObject, Type } from "./interfaces";
+import { sizeof } from "./struct-buffer";
 import {
   arrayNextProxy,
   arrayProxy,
@@ -9,7 +10,6 @@ import {
 
 export const FLOAT_TYPE = "float";
 export const DOUBLE_TYPE = "double";
-export const STRING_TYPE = "string_t";
 
 const hData: any = {
   1: {
@@ -32,7 +32,7 @@ const hData: any = {
   d: "getFloat64",
 };
 
-function typeHandle(type: StructType): [get: string, set: string] {
+function typeHandle<D, E>(type: StructType<D, E>): [get: string, set: string] {
   let h: string | undefined = undefined;
 
   const isFloat =
@@ -69,7 +69,8 @@ class StructTypeNext {
 }
 
 // D decode return type
-export class StructType extends Array<StructType> {
+// E encode obj type
+export class StructType<D, E> extends Array<StructType<D[], E[]>> {
   names: string[];
   deeps: number[] = [];
 
@@ -88,7 +89,7 @@ export class StructType extends Array<StructType> {
     return this.deeps.reduce((acc, it) => (acc *= it), 1);
   }
 
-  is(type: StructType): boolean {
+  is<D, E>(type: StructType<D, E>): boolean {
     return type.names.some((name) => this.names.includes(name));
   }
 
@@ -102,7 +103,8 @@ export class StructType extends Array<StructType> {
   constructor(
     typeName: string | string[],
     public readonly size: 1 | 2 | 4 | 8,
-    public readonly unsigned: boolean
+    public readonly unsigned: boolean,
+    private readonly KlassType?: Type<StructType<any, any>>
   ) {
     super();
 
@@ -119,7 +121,15 @@ export class StructType extends Array<StructType> {
         o.get,
         o.set
       );
-      Object.setPrototypeOf(newProxy, StructType.prototype);
+
+      // 修复: BitsType[n]
+      if (o.bits) {
+        newProxy.bits = o.bits;
+      }
+      Object.setPrototypeOf(
+        newProxy,
+        this.KlassType ? this.KlassType.prototype : StructType.prototype
+      );
       return newProxy;
     });
   }
@@ -135,38 +145,21 @@ export class StructType extends Array<StructType> {
    * @param view
    * @param littleEndian
    * @param offset
-   * @param textDecode
    */
   decode(
     view: ArrayBufferView | number[],
     littleEndian: boolean = false,
-    offset: number = 0,
-    textDecode?: TextDecoder
-  ): any {
+    offset: number = 0
+  ): D {
     view = makeDataView(view);
 
-    const isString = this.isName(STRING_TYPE);
     const result: AnyObject[] = [];
     let i = this.count;
     while (i--) {
-      let data = (view as any)[this.get](offset, littleEndian);
-      if (isString) {
-        // 截断字符串
-        if (data === 0) break;
-        if (!textDecode) textDecode = new TextDecoder();
-        data = textDecode!.decode(new Uint8Array([data]));
-      }
-      result.push(data);
+      result.push((view as any)[this.get](offset, littleEndian));
       offset += this.size;
     }
-
-    // string_t[2] => 'ab'
-    // string_t[2][1] => ['a', 'b']
-    if (isString && this.deeps.length < 2) return result.join("") as any;
-
-    return this.isList
-      ? unflattenDeep(result, this.deeps, isString)
-      : result[0];
+    return this.isList ? unflattenDeep(result, this.deeps, false) : result[0];
   }
 
   /**
@@ -184,31 +177,19 @@ export class StructType extends Array<StructType> {
    * @param littleEndian
    * @param offset
    * @param view
-   * @param textEncoder
    */
   encode(
-    obj: any /* AnyObject | string | number | undefined */,
+    obj: E,
     littleEndian: boolean = false,
     offset: number = 0,
-    view?: DataView,
-    textEncoder?: TextEncoder
+    view?: DataView
   ): DataView {
     const v = createDataView(this.count * this.size, view);
 
-    const isString = this.isName(STRING_TYPE);
-
-    if (this.isList && Array.isArray(obj)) {
-      obj = obj.flat();
-      if (isString) obj = obj.join("");
-    }
-
-    if (isString) {
-      if (!textEncoder) textEncoder = new TextEncoder();
-      obj = textEncoder.encode(obj);
-    }
+    if (this.isList && Array.isArray(obj)) (obj as any) = obj.flat();
 
     for (let i = 0; i < this.count; i++) {
-      const it = (this.isList ? obj[i] : obj) ?? 0;
+      const it = (this.isList ? (obj as any)[i] : obj) ?? 0;
       try {
         (v as any)[this.set](offset, it, littleEndian);
       } catch (error) {
@@ -221,42 +202,268 @@ export class StructType extends Array<StructType> {
   }
 }
 
-export class BitsType<T extends AnyObject> extends StructType {
-  constructor(size: 2 | 1 | 4 | 8, public readonly bits: T) {
-    super("<bits>", size, true);
+interface IBitsType {
+  [k: string]: number;
+}
+export class BitsType<D, E> extends StructType<D, E> {
+  constructor(size: 2 | 1 | 4 | 8, public readonly bits: IBitsType) {
+    super("<bits>", size, true, BitsType);
   }
 
   decode(
     view: ArrayBufferView | number[],
     littleEndian: boolean = false,
     offset: number = 0
-  ): {
-    [key in keyof T]?: 1 | 0;
-  } {
-    view = makeDataView(view);
-    const result: { [k: string]: 0 | 1 } = {};
-    let data = (view as any)[this.get](offset, littleEndian);
-    Object.entries(this.bits).forEach(([k, i]) => {
-      result[k] = ((data & (1 << i)) >> i) as 0 | 1;
-    });
-    return result as any;
+  ): D {
+    const data: number[] | number = super.decode(
+      view,
+      littleEndian,
+      offset
+    ) as any;
+    if (this.isList && Array.isArray(data)) {
+      return data.map((it) => {
+        const result: { [k: string]: 0 | 1 } = {};
+        Object.entries(this.bits).forEach(([k, i]) => {
+          result[k] = ((it & (1 << i)) >> i) as 0 | 1;
+        });
+        return result;
+      }) as any;
+    } else {
+      const result: { [k: string]: 0 | 1 } = {};
+      Object.entries(this.bits).forEach(([k, i]) => {
+        result[k] = (((data as number) & (1 << i)) >> i) as 0 | 1;
+      });
+      return result as any;
+    }
   }
 
   encode(
-    obj: {
-      [key: string]: 1 | 0;
-    },
+    obj: E,
     littleEndian: boolean = false,
     offset: number = 0,
     view?: DataView
   ): DataView {
     const v = createDataView(this.count * this.size, view);
-    let flags = 0;
-    Object.entries<number>(obj).forEach(([k, v]) => {
-      const i: number = this.bits![k];
-      if (i !== undefined) flags |= v << i;
-    });
-    (v as any)[this.set](offset, flags, littleEndian);
+
+    if (this.isList && Array.isArray(obj)) {
+      for (let i = 0; i < this.count; i++) {
+        let flags = 0;
+        Object.entries<number>(obj[i]).forEach(([k, v]) => {
+          const i: number = this.bits![k];
+          if (i !== undefined) flags |= v << i;
+        });
+        (v as any)[this.set](offset, flags, littleEndian);
+        offset += this.size;
+      }
+
+      return v;
+    } else {
+      let flags = 0;
+      Object.entries<number>(obj as any).forEach(([k, v]) => {
+        const i: number = this.bits![k];
+        if (i !== undefined) flags |= v << i;
+      });
+      (v as any)[this.set](offset, flags, littleEndian);
+      return v;
+    }
+  }
+}
+
+export class BoolType<
+  D extends boolean,
+  E extends boolean | number
+> extends StructType<D, E> {
+  constructor(typeName: string | string[], type: StructType<number, number>) {
+    super(typeName, type.size, type.unsigned, BoolType);
+  }
+
+  /**
+   * ```
+   * bool.decode([1])
+   * => true
+   *
+   * BOOL.decode([0, 0, 0, 1])
+   * => true
+   * ```
+   * @param view
+   * @param littleEndian
+   * @param offset
+   */
+  decode(
+    view: ArrayBufferView | number[],
+    littleEndian: boolean = false,
+    offset: number = 0
+  ): D {
+    let r = super.decode(view, littleEndian, offset) as any;
+    if (Array.isArray(r)) {
+      r = r.flat().map((it) => Boolean(it));
+      r = unflattenDeep(r, this.deeps);
+    } else {
+      r = Boolean(r);
+    }
+    return r;
+  }
+
+  /**
+   * ```
+   * bool.encode(0)
+   * => <00>
+   *
+   * BOOL.encode(0)
+   * => <00 00 00 00>
+   * ```
+   * @param obj
+   * @param littleEndian
+   * @param offset
+   * @param view
+   */
+  encode(
+    obj: E,
+    littleEndian: boolean = false,
+    offset: number = 0,
+    view?: DataView
+  ): DataView {
+    if (obj && Array.isArray(obj)) {
+      obj = obj.flat().map((it) => Number(Boolean(it))) as any;
+    } else if (obj) {
+      obj = Number(Boolean(obj)) as any;
+    }
+    return super.encode(obj, littleEndian, offset, view);
+  }
+}
+
+export class StringType extends StructType<string, string> {
+  constructor() {
+    super("string_t", 1, true, StringType);
+  }
+
+  /**
+   * ```
+   * string_t[2].decode([0x61, 0x62, 0, 0x63])
+   * => ab
+   * ```
+   * @param view
+   * @param littleEndian
+   * @param offset
+   * @param textDecode
+   */
+  decode(
+    view: ArrayBufferView | number[],
+    littleEndian: boolean = false,
+    offset: number = 0,
+    textDecode?: TextDecoder
+  ) {
+    view = makeDataView(view);
+
+    if (!textDecode) textDecode = new TextDecoder();
+
+    const result: AnyObject[] = [];
+    let i = this.count;
+    while (i--) {
+      let data = (view as any)[this.get](offset, littleEndian);
+      if (data === 0) break;
+      data = textDecode.decode(new Uint8Array([data]));
+      result.push(data);
+      offset += this.size;
+    }
+
+    // string_t[2] => 'ab'
+    // string_t[2][1] => ['a', 'b']
+    if (this.deeps.length < 2) return result.join("") as any;
+
+    return this.isList ? unflattenDeep(result, this.deeps, true) : result[0];
+  }
+
+  /**
+   * ```
+   * string_t[2].encode("abcd" as any)
+   * =>  <61 62>
+   * ```
+   * @param obj
+   * @param littleEndian
+   * @param offset
+   * @param view
+   * @param textEncoder
+   */
+  encode(
+    obj: string,
+    littleEndian: boolean = false,
+    offset: number = 0,
+    view?: DataView,
+    textEncoder?: TextEncoder
+  ): DataView {
+    const v = createDataView(this.count * this.size, view);
+
+    if (Array.isArray(obj)) (obj as any) = obj.flat().join("");
+
+    if (!textEncoder) textEncoder = new TextEncoder();
+    const bytes = textEncoder.encode(obj);
+
+    for (let i = 0; i < this.count; i++) {
+      const it = bytes[i] ?? 0;
+      try {
+        (v as any)[this.set](offset, it, littleEndian);
+      } catch (error) {
+        (v as any)[this.set](offset, BigInt(it), littleEndian);
+      }
+      offset += this.size;
+    }
+
+    return v;
+  }
+}
+
+export class PaddingType extends StructType<number, number> {
+  constructor() {
+    super("padding_t", 1, true, PaddingType);
+  }
+
+  /**
+   * ```
+   * padding_t[2].decode([1, 2, 3])
+   * => [ 1, 2 ]
+   * ```
+   * @param view
+   * @param littleEndian
+   * @param offset
+   */
+  decode(
+    view: ArrayBufferView | number[],
+    littleEndian: boolean = false,
+    offset: number = 0
+  ) {
+    view = makeDataView(view);
+    let i = sizeof(this);
+    const r: number[] = [];
+    while (i--) {
+      r.push((view as any)[this.get](offset, littleEndian));
+      offset++;
+    }
+    return r as any;
+  }
+
+  /**
+   *
+   * ```
+   * padding_t[10].encode(0 as any)
+   * => <00 00 00 00 00 00 00 00 00 00>
+   * ```
+   *
+   * @param zero
+   * @param littleEndian
+   * @param offset
+   * @param view
+   */
+  encode(
+    zero: number = 0,
+    littleEndian: boolean = false,
+    offset: number = 0,
+    view?: DataView
+  ): DataView {
+    const v = createDataView(this.count * this.size, view);
+    if (typeof zero !== "number") zero = 0;
+    let length = sizeof(this);
+    while (length-- > 0) v.setUint8(offset++, zero);
     return v;
   }
 }
@@ -272,12 +479,12 @@ export class BitsType<T extends AnyObject> extends StructType {
  * @param size
  * @param unsigned
  */
-export function registerType(
+export function registerType<D extends number, E extends number>(
   typeName: string | string[],
   size: 1 | 2 | 4 | 8,
   unsigned = true
-): StructType {
-  return new StructType(typeName, size, unsigned);
+) {
+  return new StructType<D, E>(typeName, size, unsigned);
 }
 
 /**
@@ -290,18 +497,21 @@ export function registerType(
  * @param typeName
  * @param type
  */
-export function typedef(typeName: string | string[], type: StructType) {
-  const newType = registerType(typeName, type.size, type.unsigned);
+export function typedef<D extends number, E extends number>(
+  typeName: string | string[],
+  type: StructType<any, any>
+) {
+  const newType = registerType<D, E>(typeName, type.size, type.unsigned);
   return newType;
 }
 
-interface IBitsType {
-  [k: string]: number;
-}
-
-export function bits<T extends IBitsType>(
-  type: StructType,
-  obj: T
-): BitsType<T> {
-  return new BitsType<T>(type.size, obj);
+export function bits<T extends IBitsType>(type: StructType<any, any>, obj: T) {
+  return new BitsType<
+    {
+      [key in keyof T]: 1 | 0;
+    },
+    {
+      [key in keyof T]?: 1 | 0;
+    }
+  >(type.size, obj);
 }