+++ src/index.ts
@@ -3,3 +3,5 @@ export * from "./types";
 export { registerType, typedef, bits } from "./class-type";
 export * as CStruct from "./c-struct";
 export { display } from "./display";
+export { createDataView, sbytes, sview } from "./utils";
+export { pack, unpack, calcsize } from "./py-struct";